% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormalReturn.R
\name{abnormalReturn}
\alias{abnormalReturn}
\title{Calculation of abnormal returns}
\usage{
abnormalReturn(prices_stock, prices_market = NULL, from = "2015-01-01",
  to = "2015-12-31", eventIndex = NULL, model = "marketmodel",
  estimationWindowLength = 10, c = 10, attributeOfInterest = "Close",
  showPlot = FALSE)
}
\arguments{
\item{prices_stock, prices_market}{Data frames, minimally containing a column
\code{Date} and a second column named as specified by
\code{attributeOfInterest}. Time series data of prices_market and prices_stock
performance respectively. \strong{OR:} Character objects, representing stock symbols.
These symbols are used to query data directly from the Yahoo Finance platform.}

\item{from, to}{Character objects. Defining the data range.}

\item{eventIndex}{Positive integer, sequence of integers, or \code{NULL}.
If \code{eventIndex} is a positive integer or sequence of integers, function computes
abnormalReturns for the defined indices only. If \code{eventIndex=NULL}, function
computes abnormalReturns for all data points in the range of
\code{eventWindowLength+1:nrow{prices_market}}.}

\item{model}{A character object.
\itemize{
     \item \strong{\code{"marketmodel"}}
     \item \strong{\code{"constantmeanmodel"}}
     Argument 'prices_market' is not required for the constant mean model
   }
   All other types give an error.}

\item{estimationWindowLength}{Positive integer. Defines number of
observations considered to estimate model(s).}

\item{c}{Positive integer. Cumulative Abnormal Returns, defines the number of abnormal returns to cumulate.}

\item{attributeOfInterest}{A character object. Defines the attribute of
interest.}

\item{showPlot}{A boolean value. Should a plot of the prices_stock performance
be shown?}
}
\value{
\code{abnormalReturn} returns a data frame, comprising following columns:
  \itemize{
     \item \strong{\code{Date}} POSIXct.
     \item \strong{\code{abnormalReturn}} Numerical.
     \item \strong{\code{cumulativeAbnormalReturn}} Numerical. The first \code{c-1} rows will be \code{cumulativeAbnormalReturn=NA}.
     \item \strong{\code{stockReturn}} Numerical. Actual return of prices_stock.
   }
   The number of rows returned depends on the length of \code{prices_stock}/\code{prices_market},
   as well as \code{estimationWindowLength} and \code{eventIndex}.
}
\description{
This functions implements the event study
  methodology and abnormal returns in particular. The event study methodology
  is a common way to study the effects of certain events on stock prices. It
  thus calculates a so-called abnormal return that measures the impact
  without confounding influences. As part of this method, one first has to
  predict a normal return in the absence of the event under study.
  Afterwards, one calculates the difference between the actual return and the
  previously predicted normal return, i.e. the abnormal return.
}
\examples{
x <- abnormalReturn(prices_stock=d.VW, prices_market=d.DAX, model="marketmodel",
                    estimationWindowLength=10, c=10,
                    attributeOfInterest="Close", showPlot=TRUE)
head(x)
summary(x$abnormalReturn)
}
\keyword{abnormalReturn}
\keyword{event}
\keyword{financial}
\keyword{impact}
\keyword{markets}
\keyword{performance}
\keyword{prices_market}
\keyword{prices_stock}
\keyword{stock}
\keyword{values}

