% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormalReturn.R
\name{computeAbnormalReturn}
\alias{computeAbnormalReturn}
\title{computeAbnormalReturns function}
\usage{
computeAbnormalReturn(portfolio, commodity, regressionType = "OLS",
  eventIndex = NULL, estimationWindowLength = 20,
  attributeOfInterest = "Close", showPlot = FALSE)
}
\arguments{
\item{portfolio}{Time series data of portfolio performance}

\item{commodity}{Time series data of commodity performance}

\item{regressionType}{Default: 'OLS'}

\item{eventIndex}{Index or Indexrange of desired abnormalReturns}

\item{estimationWindowLength}{Defines number of observations considered to estimate model.}

\item{attributeOfInterest}{Defines the attribute of Interest. Default: Daily 'Close'}

\item{showPlot}{Plot commodity performance. Default: FALSE}
}
\description{
Computes the abnormal returns.
}
\examples{
computeAbnormalReturn(portfolio=d.DAX, commodity=d.VW, regressionType='OLS', eventIndex=NULL,
                      estimationWindowLength=20, attributeOfInterest='Close', showPlot=TRUE)
}
\keyword{(ToDo)}
\keyword{cats}

