% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.Adidas}
\alias{d.Adidas}
\title{Exemplary Commodity data set: Adidas 2015}
\format{A data frame with 253 rows and 7 variables:
\describe{

  \item{Adj.Close}{Daily Adj.Close}
  \item{Close}{Daily Close}
  \item{Date}{Date}
  \item{High}{Daily High}
  \item{Low}{Daily Low}
  \item{Open}{Daily Open}
  \item{Volume}{Daily Volume}
}}
\source{
\url{https://de.finance.yahoo.com/q/hp?s=ADS.DE}
}
\usage{
d.Adidas
}
\description{
A dataset containing the daily details of the Adidas stock 2015. Adidas is listed in the DAX portfolio.
}
\keyword{datasets}

