% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.DAX}
\alias{d.DAX}
\title{Exemplary Portfolio data set: DAX 2015}
\format{A data frame with 253 rows and 7 variables:
\describe{

  \item{Adj.Close}{Daily Adj.Close}
  \item{Close}{Daily Close}
  \item{Date}{Date}
  \item{High}{Daily High}
  \item{Low}{Daily Low}
  \item{Open}{Daily Open}
  \item{Volume}{Daily Volume}
}}
\usage{
d.DAX
}
\description{
A dataset containing the daily details of the DAX 2015. The DAX portfolio list, among others, Adidas and VW stocks.
}
\keyword{datasets}

