% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnormalReturn.R
\name{computeAbnormalReturn}
\alias{computeAbnormalReturn}
\title{computeAbnormalReturns function}
\usage{
computeAbnormalReturn(data = NULL, datasetname = NULL, orderBy = "Date",
  showPlot = TRUE, attributeOfInterest = "Close", eventWindowLength = 1,
  eventWindowStart = NULL, estimationWindowLength = NULL, summary = FALSE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function computes the abnormal returns you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

