% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.DAX}
\alias{d.DAX}
\title{Exemplary portfolio data set: DAX 2015}
\format{A data frame with 253 rows and 7 variables:
\describe{
  \item{Adj.Close}{Daily Adj.Close}
  \item{Close}{Daily Close}
  \item{Date}{Date}
  \item{High}{Daily High}
  \item{Low}{Daily Low}
  \item{Open}{Daily Open}
  \item{Volume}{Daily Volume}
}}
\source{
\url{https://de.finance.yahoo.com/q/hp?s=ADS.DE}
}
\usage{
d.DAX
}
\description{
A dataset containing the daily ticks of the DAX 2015.
}
\keyword{datasets}

